+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Elasticsearch plugin for Craft CMS 3.x Changelog
 
+## 1.1.0 - Unreleased
+### Added
+- Craft Commerce product support
+
+### Changed
+- IndexEntryException has been replaced by IndexElementException
+
 ## 1.0.0 - 2018-12-12
 ### Added
 - Initial release